requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
boto3==1.28.12
botocore==1.31.12
CacheControl==0.14.0
cachetools==5.3.3
celery==5.3.6
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.38.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.7
cron-descriptor==1.4.3
cryptography==41.0.3
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.4
django-allauth==0.54.0
django-celery-beat==2.6.0
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==23.2
django-js-asset==2.2.0
django-oauth-toolkit==3.0.1
django-oauth2-provider==0.2.6.1
django-rest-auth==0.9.5
django-rest-passwordreset==1.3.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==3.1.0
drf-standardized-errors==0.14.1
drf-yasg==1.21.6
facebook-sdk==3.1.0
firebase-admin==6.5.0
google-api-core==2.18.0
google-api-python-client==2.124.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jwcrypto==1.5.6
kombu==5.3.6
Markdown==3.4.3
MarkupSafe==2.1.3
msgpack==1.0.8
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pluggy==1.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
PyJWT==2.8.0
pyotp==2.8.0
pyparsing==3.1.2
pytest==7.3.1
pytest-cov==4.0.0
pytest-django==4.5.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-instagram==1.3.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
setuptools==75.2.0
shortuuid==1.0.13
simplejson==3.19.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
tweepy==4.14.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.16
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.2.0
